requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-auth==1.19.2
google-auth-oauthlib==0.4.1
gspread==3.6.0
httplib2==0.18.1
idna==2.10
lxml==4.5.2
oauth2client==4.1.3
oauthlib==3.1.0
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
